Implement sexp#repeat()
We use a script-level variable s:repeat instead of parametrizing the
repeat level since only a few text objects will need to respond to a
repeat count. s:repeat allows us to simply wrap the functions we want to
repeat, and the functions themselves can handle s:repeat if desired.

Vim is profoundly single-threaded, so there is no reason to worry about
concurrency issues. However, if we find that we need another such
implicit variable in the future, we will revisit this decision.