Remove loop from s:nearest_bracket for large speed boost
As discussed in the comments, this change also will treat mismatched
brackets as the same type. The following, for example, will select fine:

    (defn [foo)
      {bar quux])

What we get in return is a massive speed up in some pathological cases.
For instance, selecting 7 nested forms via

    v7af

in a function in Clojure's core.clj took ~9.647 seconds previously, and
about ~0.244 seconds after the change, for a 40x speedup!

Anticipating an analogous speed up in the other searchpairpos calls, we
change the default g:sexp_maxlines to 0.