Use control codes to mark displaced elements
After wrestling with various solutions, it became clear that the
arithmetic solution to finding the terminals of the original selection
after swapping forward required at least 4 different branches.

Using ASCII START OF TEXT and END OF TEXT control characters to delimit
the next element is a quite simple and hackish solution. If we find that
there are many source files with literal 0x02 and 0x03 characters, we
can delimit with chains of control characters.