Refine defplug to correctly pass v:count to repeat#set
This requires prepending the main function call, so we just pass a
callable expression to defplug.

Hiding the actual definition of the <Plug> mappings in a function
wrapper was something I wanted to avoid, but this is really the only
clean way to accomplish vim-repeat integration without lines 200+
characters long.