Rename sexp_lift_* to sexp_raise_*
This was the original name for mapping/function, but I changed it when I
was led to believe that paredit's sexp raise did not do the same.

After installing paredit.el and trying it out for myself, it appears
to be essentially the same as sexp_lift_element, so we are changing it
back!