Fall back to current element when selecting compound forms
This breaks from the behavior of the builtin bracket text objects, which
do not fallback to the current word, but is strictly more useful.

Lisp forms are defined in terms of their ability to be evaluated, so
falling back to an atomic form when a compound one does not exist is
unsurprising.

Suggested by Tim Pope.

Closes #2