Refactored s:defplug() to fix various bugs and made vim-sexp work with smear-cursor.
Some of the bugs involved the incorrect handling of counts in commands that support them.
There was also an issue caused by use of the pattern '\v%(<call>)@=', which did not work
as intended when the 'iskeyword' option contained '>' (as it may in some Lisp buffers).
This issue resulted in malformed command maps, which raised errors upon invocation.

The original impetus for this change was that vim-sexp didn't play well with the
smear-cursor plugin. The root of the problem was that vim-sexp used the `:<c-u>' idiom
(now obsolete) to introduce maps, an approach which generated *lots* of 'CmdlineChanged'
autocmd events. Smear-cursor's cursor animation logic assumes 'CmdlineChanged' indicates
the user is typing on the command line, with the result that vim-sexp commands had either
inconsistent or nonexistent animation. The fix, for Vim/Neovim instances that support it,
is to use the <cmd> modifier in lieu of the problematic `:<c-u>'.

Note: While testing, discovered that the 'which-key' plugin can swallow the count
associated with the <leader> keystroke in such a way as to break usage of a count with
vim-sexp maps that begin with <leader>. In my case, this meant that the clone command
ignored a count until I disabled 'which-key'. Not sure what the best workaround for this
is, other than metafying sexp maps to avoid <leader>.

Note: The DEFPLUG command is no longer used to define any commands, and should probably
be removed at some point, with corresponding simplifications applied to s:defplug().