Revert logic that treats null pos from s:super_range() as error.
The logic might make sense once treesitter support is implemented, but
for now, a treesitter user who hasn't enabled legacy syntax regions
would likely find it annoying, as the failure to detect comments tends
to produce lots of spurious errors. Better to select a small portion of
a comment than to pop up an error every time user attempts to do so.