Backup of work in progress on treesitter support.
At this point, much of the framework is in place, but lots of test and
cleanup are needed, and there are likely still significant bugs. I also
haven't determined whether the caching provides sufficent value to
justify keeping it, now that I've reworked the higher level logic to
require far fewer treesitter tree queries.