Another backup of work in progress.
TODO: Just noticed I can skip a lot of searching for containing brackets
in the very common scenario of an inner element selection initiated with
a single character selected (vie). Currently, the call to super_range()
is unconditional, but it's unnecessary when only one character is
selected, or for that matter, when both ends of the selection are within
the same TSNode.