Refactored set_marks_around_current_element() to remove statefulness.
Background: Originally, used heuristics of various sorts to determine
whether ie/ae should expand or just cleanup current selection. Early on,
was using is_dirty(), which was intended to detect repeat invocations of
the same command. Dissatisfied with that approach (partly because of its
complexity/fragility), went to an approach that considered only whether
the ie/ae would *change* the current selection. One problem with that
approach is that it can make it difficult for the user to predict what
an invocation of ie/ae will do; another problem is that it creates a
corner case that violates the POLS... When cursor was on a single-char
atom (in normal mode), hitting vie resulted in the selection of *two*
elements. Though somewhat disconcerting (to me, at least), this behavior
was actually "correct", according to the heuristic employed: since the
single-char atom was already completely inner-selected, inner-selecting
it changed nothing, and thus did not consume a count, thereby allowing a
second element to be pulled in.

Both because I don't like violating the principle of least surprise, and
because the complexity added by the stateful approaches didn't seem to
be sufficiently offset by any corresponding benefit, I've removed all
statefulness from set_marks_around_current_element(), adopting the
following approach to extending selection... Never expand selection
unless an explicit count is used. With an explicit count, the 1st count
refers to the *current* selection, with only [count-1] additional
elements pulled into the selection. In other words, the first count is
always for completing or cleaning up whatever is currently selected, and
[count] represents the total number of "elements" that should be
selected after the operation completes, with the initial selection
(partial or full) counting as a single "element".