Backing up progress on the outer element selection adjustment refactor.
I've implemented (but not yet documented) several new options that allow
the user to customize the selection of leading/trailing whitespace on an
"outer element" selection. I like the new logic *much* better than the
old, and I think it does a much better job of adhering to "the principle
of least surprise". Also, the implementation is getting simpler.

Lots of cleanup still needed, but it's been awhile since I backed up...