Fixed issue with 'textwidth' / overrun criterion cost calculation.
Prior to this, only a group's worst-case overrun was considered. This
was problematic because it meant that once a group had incurred an
overrun cost, the algorithm sought to amortize it by pulling additional
comment lines into the group, even if the additional lines had as much
overrun as the worst-case! Not good. The fix for now is to consider not
only worst case, but also *average* overrun scaled by number of comments
in the group.

Also, tweaked a few default option values.

TODO: Grouping criteria default weights may warrant additional tweaking,
but at this point, I'm seeing mostly unsurprising behavior.