Fixed Tresitter-specific bug in detection of top-level.
The bug manifested as sexp#clone sometimes performing a single-line
clone for top-level elements, for which a multi-line clone should have
been selected.
Root Cause: The lua nearest_bracket() function contained a loop that
traversed the tree upwards looking for a node that begins with a
bracket. The loop test contained an implicit, incorrect assumption:
namely, that all tree nodes (including the root node) represent
syntactic elements in the file, such as lists and elements. This logic
failed to account for the fact that all source files contain a root
node, typically with a name like "source" or "program", representing the
entire file, and not any specific, user-visible element within it.
Because the root "source" node begins at (0,0), an open bracket at the
first position in the file would mistakenly be treated as the bracket
opening a nonexistent list containing the entire file. Thus, whenever
there was an open bracket in line 1, col 1, the s:at_top() call (which
relied on ts.nearest_bracket() whenever legacy syntax was not in use)
concluded that the cursor was within a list, even when it wasn't. As a
consequence, s:at_top() would invariably return false whenever the first
char in the file was an open bracket.
Fix: Ignore the root node when searching for containing forms.