Fixed bugs affecting multi-line clone.
There were actually 2 distinct bugs here:
The first problem was that the call to select_current_element() in
get_clone_target_range() was not followed by an escape to normal mode,
with the result that we remained in visual mode until the point at which
yankdel_range() selected and yanked text. This was a problem because
set_visual_marks() didn't work properly in visual mode. In particular,
you can't just set the '< and '> marks with setpos() and expect the
visual selection to be immediately updated. To summarize the problematic
sequence...
1. select_current_element left us in visual mode
2. a subsequent cursor() call from an unrelated function made an
   unintended adjustment to the visual selection (again, leaving us in
   visual mode).
3. yankdel_range()'s call to set_visual_marks() changed the marks, but
   not the actual visual selection that was used for its yank command,
   with the result that the wrong text was yanked (and therefore cloned).

Fix: First of all, added the missing escape to Normal mode after the
call to select_current_element(). Note that in the absence of that bug,
the flaw with set_visual_marks() wouldn't have been noticed, since we
wouldn't have been in visual mode when the call to it was made. Also,
rewrote set_visual_marks() so that it now works in both visual and
normal modes.
TODO: Consider whether it wouldn't make more sense just to use the
legacy version of set_visual_marks() unconditionally, given that the
simpler approach is invalid, at least in visual mode.