Added docount_stateful() and made capture/emit support auto-reindent.
The auto-reindent is conditional on newly-added option
g:sexp_capture_emit_does_indent, which is set by default because
auto-indent just makes capture/emit "feel" so much nicer.
Rationale: While it's true that the user could perform cleanup manually
(potentially after a sequence of captures/emits is complete), by making
it easier to visualize the results of the operation, auto-reindent
reduces the risk of error (e.g., slurping too many or too few elements).

dcount_stateful() was added to address suboptimal aspects of the current
implementation, especially in the area of exception handling. Under the
current implementation, an exception thrown during an iteration did not
terminate docount(), which continued invoking the operation function for
the remaining counts. E.g., even if the first iteration of a capture
with [count] == 1M realized there were no elements to capture, docount()
would stilll call stackop_capture() an additional 999K times. A possibly
more significant issue with the existing approach was that it provided
no way to perform cleanup/finalization at the *end* of a counted
sequence of operations. This sort of cleanup was needed to support the
new capture/emit auto-indent feature: because re-indent can trigger
whitespace cleanup and/or comment alignment, it shouldn't be done in the
middle of [count] capture/emit iterations, but only at the very end.
Note: Although docount_stateful() was added for capture/emit
auto-indent, it is sufficiently flexible to be used by any feature
implementing the requisite __init()/__update()/__final() functions.

TODO: Consider adding auto-indent for other commands that modify forms.