Added redundant safety-check (and failsafe override flag) to yankdel_range().
Prior to this, if a bug resulted in a bad range being passed to
yankdel_range(), non-whitespace text could be deleted, even in cases
where the explicit intent of the call to yankdel_range() was whitespace
modification. Currently, all calls to yankdel_range() are intended to
delete only whitespace; thus, at the cost of a small amount of execution
time (spent in range_has_non_ws()), we can dramatically reduce the
impact of a bug that impacts the range passed to yankdel_range(). If the
optional failsafe_override flag (default 0) is not provided, any request
to modify non-whitespace will be rejected with a warning.

TODO: Figure out a good way to ensure the warning messages are never
clobbered by a subsequent redraw before user has a chance to see them.