Fixes #38 and #39: Bugs and idiosyncrasies pertaining to sexp#raise.
Added logic to make sexp_raise a NOP when the target of the raise has no
parent list. Prior to this, the futile attempts to raise tended to
produce strange and unexpected results.

When sexp_raise is invoked in visual mode, first expand selection as
necessary to construct a "super range" (i.e., a sequence of one or more
complete elements at the same level), then use that as the raise target.
Prior to this, the blind attempts to raise whatever was selected could
produce unbalanced forms.