Fix for guns/vim-sexp#10: Capturing elements like emacs slurp
This commit addresses the following usability issues, as reported by users in
the referenced Github issue:
1. Capture/emit position cursor on list bracket, requiring user to reposition
   manually (e.g., with <C-o> or `g;`) after each capture/emit before editing
   can be resumed.
2. Capture/emit fail when the containing list has no elements to capture or
   the current list has no elements to emit, respectively.

Note: Automatic re-indent of lines affected by capture/emit was implemented in
a previous commit.

Users have expressed a strong preference for Paredit-style capture/emit.
Accordingly, this commit makes the capture/emit logic more like Paredit's
(though not identical to it, as Paredit's behavior in at least one corner case
strikes me as counter intuitive, and likely more implementation than
specification-driven).

The behavior implemented in this commit can be described roughly as follows...

If a capture/emit cannot be performed on the current list, move up the sexp
tree and retry, recursively. Note that when a capture has a count, each
iteration starts from the *original* cursor position, not the position reached
by the preceding iteration. This approach ensures that elements in containing
lists can (eventually) be pulled into the current list by successive captures,
even when their containing lists are many levels up. An unfortunate
consequence of this approach is that captures with large counts within deeply
nested forms will be less efficient than they could be if we dispensed with
the `docount` wrapper and used the count to calculate all bracket relocations
up-front. I've partially implemented an approach which does just that, but
left it as work-in-progress on a branch, pending a decision on whether it
constitutes premature optimization.

For normal mode capture/emit, preserve cursor position unless option
g:sexp_never_emit_cursor is set and cursor position would be emitted; in that
case, put cursor on the bracket at the emitting end of the list doing the
emitting. (Note that Paredit seems always to emit the cursor when the element
under it is emitted.)

For visual mode capture/emit, visually select the list doing the
capturing/emitting. (Note that selecting the list effectively prevents
cursor emission.)

TODO: It occurs to me that some sort of cursor preservation logic should
probably be added for swap_list_{forwards,backwards} as well.