Fix for issue #43: vaf not working with anon functions
Forms with leading macro chars (e.g., `#( ... )`) were not being
recognized as lists when Treesitter was used.
Root Cause: Failure to position *on* macro chars before calling legacy
function to find the end of the macro char sequence.

Note: This is intended as a stopgap measure, as I'd really like to
remove the call to sexp#current_macro_character_terminal() altogether,
most likely replacing it with a pattern search of the buffer text
directly.