Make macro characters configurable
vim-sexp currently hard codes the macro characters that it understands.
This means that if a user-defined Lisp uses different macro characters
to the default Lisp (i.e. Scheme), then features of vim-sexp that depend
on macro characters will not work correctly.

An alternative approach is to allow a user to configure the macro
characters, similar to how a user can configure supported filetypes. To
achieve this, the script-local variable `s:macro_filetype_characters` is
changed to the global variable `g:sexp_filetype_macro_characters`.  The
name was selected to be consistent with `g:sexp_filetypes`.