Fixed bugs involving optimized re-indent after put and changed option format.
The logic that determined whether or not to include an extra element past the
put text was incorrectly skipped when the inserted separator was not a NL.

The logic that determined exclusive mode at end of splice did not handle all
cases.

Changed format of sexp_auto_indent_range option value from overly complex
enumeration to simpler 1 or 2 char string of single-character flags.

Issue #32