Refactored sexp#put() pipeline and fixed various logic errors.
Original pipeline was built around explicit tgt/adj locations (where the
relative order of tgt and adj was determined by the put direction). The
refactored code maintains the concept of tgt/adj, but begins much
earlier working with a simple start/end range. With the new approach,
differentiating between tgt and adj is as simple as using the
appropriate direction-based index: i.e., !tail for tgt, tail for adj.

While refactoring, simplified the logic used to determine separators and
made it more consistent. Also, fixed various logic errors, mostly
involving the handling of comments in registers.
Note: The rules still need a bit of fine-tuning: e.g., I'm thinking of
having the 2nd element of lists treated specially by the logic that
computes separators.

Added special case that makes p and P behave slightly differently when
cursor is in whitespace adjacent to (and colinear with) target. In this
special case, putting _away from_ the element inserts a NL even when
putting towards the target would not.
Issue guns#32