Just backing up a prototype version of replace mode put.
May eventually merge this, but I'm not happy with the way it handles
leading/trailing whitespace. In particular, it needs logic along the
lines of what's in terminals_with_whitespace() to avoid eating up so
much leading whitespace that user is disoriented. It also probably needs
to consider the register contents so that it can get the
leading/trailing separators right. One possibility would be to handle in
several steps: i.e., 1) outer element selection; 2) delete; 3) smart
put. But there are details that would need working out: e.g., which type
of put (before/after) and with which element as target.
I'm keeping this off the main guns#32 branch both because I haven't
performed a lot of testing on it, and also because I may not keep it,
and certainly won't keep it in its current form.